% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_db_upload.R
\name{confirm_db_upload}
\alias{confirm_db_upload}
\title{Confirm database upload}
\usage{
confirm_db_upload(con, analysis_lut)
}
\arguments{
\item{con}{Connection to... SOMETHING. Evan or colt: help!}

\item{analysis_lut}{List? Something? What options make sense?}
}
\value{
Evan or Colt, help:
}
\description{
This function attempts to verify that the model estimates have been written as expected and that the same number of
rows, columns, and values have been written. Currently this is in draft form and needs to be more comprehensive.
}
\details{
Dev note: I assume this should be exported. If so, we may want to add some input validation
}
