% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreelEstimateR-package.R
\docType{package}
\name{CreelEstimateR-package}
\alias{CreelEstimateR}
\alias{CreelEstimateR-package}
\title{CreelEstimateR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cbedwards-dfw/CreelEstimateR}
  \item \url{https://cbedwards-dfw.github.io/CreelEstimateR/}
  \item Report bugs at \url{https://github.com/cbedwards-dfw/CreelEstimateR/issues}
}

}
\author{
\strong{Maintainer}: Collin Edwards \email{Collin.Edwards@dfw.wa.gov} (\href{https://orcid.org/0000-0002-4937-5159}{ORCID})

}
\keyword{internal}
