% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_pe_catch.R
\name{est_pe_catch}
\alias{est_pe_catch}
\title{Title}
\usage{
est_pe_catch(days, pe_inputs_list, ...)
}
\arguments{
\item{days}{Help!}

\item{pe_inputs_list}{Help!}

\item{...}{Why do we have this?}
}
\value{
Help!
}
\description{
Dev notes: unclear why we include \code{...}.
Dev notes: missing \code{params}, \code{dwg}.
}
