% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_db_con.R
\name{establish_db_con}
\alias{establish_db_con}
\title{Establish database connection}
\usage{
establish_db_con(max_attempts = 5, delay_seconds = 3)
}
\arguments{
\item{max_attempts}{Numeric. Number of attempts to make when attempting to establish connection. Defaults to 5.}

\item{delay_seconds}{Numeric. Number of seconds to wait between attempt. Defaults to 3.}
}
\value{
Connection to WDFW PostgreSQL database.
}
\description{
Establishes a connection with the WDFW Postgres database using local configuration and user credentials.
}
\details{
requires a local config file that is on gitignore list
Dev note: we could add template config file to the inst/ folder, and then add
a function that sets up the config file.
}
