% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrity.R
\name{fetch_fishery_names}
\alias{fetch_fishery_names}
\title{List all available fishery names}
\usage{
fetch_fishery_names()
}
\value{
Character vector
}
\description{
List all available fishery names
}
\examples{
head(fetch_fishery_names())
}
