% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bss.R
\name{fit_bss}
\alias{fit_bss}
\title{Wrapper on \code{stan()}}
\usage{
fit_bss(
  model_file_name = here::here("stan_models/BSS_creel_model_02_2024-04-03.stan"),
  bss_inputs_list,
  n_chain = n_chain,
  n_cores = n_cores,
  n_iter = n_iter,
  n_warmup = n_warmup,
  n_thin = n_thin,
  adapt_delta = adapt_delta,
  max_treedepth = max_treedepth,
  init = "0",
  ...
)
}
\arguments{
\item{model_file_name}{??}

\item{bss_inputs_list}{??}

\item{n_chain}{??}

\item{n_cores}{??}

\item{n_iter}{??}

\item{n_warmup}{??}

\item{n_thin}{??}

\item{adapt_delta}{??}

\item{max_treedepth}{??}

\item{init}{??}

\item{...}{??}
}
\value{
Results of stan run
}
\description{
Dev note: we don't want our default arguments to be based on global variables.
}
