% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_analysis_lut.R
\name{generate_analysis_lut}
\alias{generate_analysis_lut}
\title{Generate base analysis look up table with session-specific uuid}
\usage{
generate_analysis_lut(params)
}
\arguments{
\item{params}{??}
}
\value{
??
}
\description{
Dev note: For a package, we really want to avoid using \code{assign}. Can we just return this?
}
