% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bss_cpue_daily.R
\name{get_bss_cpue_daily}
\alias{get_bss_cpue_daily}
\title{get daily bss cpue}
\usage{
get_bss_cpue_daily(bss_fit, ecg, ...)
}
\arguments{
\item{bss_fit}{??}

\item{ecg}{??}

\item{...}{??}
}
\value{
??
}
\description{
get daily bss cpue
}
