% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fishery_data.R
\name{get_fishery_data}
\alias{get_fishery_data}
\title{Obtain and format interview, catch, and effort data for sets of fishery-years}
\usage{
get_fishery_data(fishery_name, years)
}
\arguments{
\item{fishery_name}{Character string of fishery name. If a vector of character strings is provided instead, will use all combinations of fishery_name by years.}

\item{years}{Integer or vector of integers identifying years of data to pull}
}
\value{
List with three dataframes: \verb{$interview}, \verb{$catch}, and \verb{$effort}
}
\description{
Based on code written for descriptive_statistics.R in the \code{creelreview} package. Streamlines
the process of getting creel data in useful format.
}
\examples{
\dontrun{
temp <- get_fishery_data(
  fishery = "Nisqually salmon",
  year = 2021:2023
)
}
}
