% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_est_pe_catch.R
\name{plot_est_pe_catch}
\alias{plot_est_pe_catch}
\title{Title}
\usage{
plot_est_pe_catch(estimates_pe_catch, est_catch_group, period_pe, ...)
}
\arguments{
\item{estimates_pe_catch}{??}

\item{est_catch_group}{??}

\item{period_pe}{??}

\item{...}{??}
}
\value{
??
}
\description{
Title
}
