% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inputs_pe_census_vs_index.R
\name{plot_inputs_pe_census_vs_index}
\alias{plot_inputs_pe_census_vs_index}
\title{Plot paired season-long counts from census and index angler effort count surveys}
\usage{
plot_inputs_pe_census_vs_index(census_TI_expan)
}
\arguments{
\item{census_TI_expan}{??}
}
\value{
??
}
\description{
Plot paired season-long counts from census and index angler effort count surveys
}
