% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inputs_pe_cpue_period.R
\name{plot_inputs_pe_cpue_period}
\alias{plot_inputs_pe_cpue_period}
\title{Plot cpue by period, section_num, angler_final, and catch group}
\usage{
plot_inputs_pe_cpue_period(
  dwg,
  days,
  dwg_summarized,
  daily_cpue_catch_est,
  est_catch_group,
  period_pe
)
}
\arguments{
\item{dwg}{??}

\item{days}{??}

\item{dwg_summarized}{??}

\item{daily_cpue_catch_est}{??}

\item{est_catch_group}{??}

\item{period_pe}{??}
}
\value{
??
}
\description{
aggregating over day_type to simplify plot
}
