% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inputs_pe_index_effort_counts.R
\name{plot_inputs_pe_index_effort_counts}
\alias{plot_inputs_pe_index_effort_counts}
\title{Plot paired season-long counts from census and index angler effort count surveys}
\usage{
plot_inputs_pe_index_effort_counts(effort_index)
}
\arguments{
\item{effort_index}{??}
}
\value{
??
}
\description{
Dev note: initial version started pipe with \code{dwg_summ$effort_index}, did not use argument.
I'm assuming argument is supposed to be used instead. May be wrong.
}
