% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_days.R
\name{prep_days}
\alias{prep_days}
\title{Title}
\usage{
prep_days(
  date_begin,
  date_end,
  weekends = c("Saturday", "Sunday"),
  holidays,
  lat,
  long,
  period_pe,
  sections,
  closures,
  day_length,
  day_length_inputs
)
}
\arguments{
\item{date_begin}{??}

\item{date_end}{??}

\item{weekends}{??}

\item{holidays}{??}

\item{lat}{??}

\item{long}{??}

\item{period_pe}{??}

\item{sections}{??}

\item{closures}{??}

\item{day_length}{??}

\item{day_length_inputs}{??}
}
\value{
??
}
\description{
Title
}
