% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dwg_census_expan.R
\name{prep_dwg_census_expan}
\alias{prep_dwg_census_expan}
\title{creates summary table of census expansion factors based on values in data base}
\usage{
prep_dwg_census_expan(eff)
}
\arguments{
\item{effort}{data from dwg filtered using start & end dates passed from params}
}
\value{
??
}
\description{
NOTE: "p_census is the same as "p_TI" in the BSS model, which represent a known or assumed proportion of each section covered by tie-in counts for each final angler-type; this parameter is used to directly expand/modify the census count data prior to the estimate of the bias parameter/expansion factor
NOTE: values of "p_census/p_TI" are entered and stored in creel db
NOTE: currently, values of "p_census/p_TI" are limited to two "angler_final" groupings (bank, boat)
}
