% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dwg_effort_index.R
\name{prep_dwg_effort_index}
\alias{prep_dwg_effort_index}
\title{Aggregates index effort counts over locations within count_seq & section based on study_design and user input values for boat_type_collapse, fish_location_determines_type, angler_type_kayak_pontoon}
\usage{
prep_dwg_effort_index(
  eff,
  study_design,
  boat_type_collapse = NA,
  fish_location_determines_type = NA,
  angler_type_kayak_pontoon = NA,
  params
)
}
\arguments{
\item{eff}{effort data from dwg filtered using start & end dates passed from params}

\item{study_design}{string passed from params denoting which study design was followed during data collection}

\item{boat_type_collapse}{string passed from params that controls whether all (potential) boat types (e.g., motor_boat, drift_boat) are collapsed (i.e., boat_type_collapse: "Yes") into a single boat type or kept separate (boat_type_collapse: "No").}

\item{fish_location_determines_type}{string passed from params that controls whether the observed fishing location for a given angler group during an effort count determines their angler type.}

\item{angler_type_kayak_pontoon}{string passed from params that controls whether a boat designated as a kayak, pontoon, or kick during an effort count or angler group interview should be designated as a boat or bank angler.}

\item{params}{??}
}
\value{
?
}
\description{
Note dplyr::summarize() does not account for missed locations with date-section-sequence
}
