% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dwg_interview_angler_types.R
\name{prep_dwg_interview_angler_types}
\alias{prep_dwg_interview_angler_types}
\title{Create angler_final field based on study_design designation and user input values for boat_type_collapse, fish_location_determines_type, angler_type_kayak_pontoon}
\usage{
prep_dwg_interview_angler_types(
  interview_fishing_time,
  study_design,
  boat_type_collapse = NA,
  fish_location_determines_type = NA,
  angler_type_kayak_pontoon = NA
)
}
\arguments{
\item{interview_fishing_time}{output from preceding function that calculates fishing time of angler groups}

\item{study_design}{string passed from params denoting which study design was followed during data collection}

\item{boat_type_collapse}{string passed from params that controls whether all (potential) boat types (e.g., motor_boat, drift_boat) are collapsed (i.e., boat_type_collapse: "Yes") into a single boat type or kept separate (boat_type_collapse: "No").}

\item{fish_location_determines_type}{string passed from params that controls whether the observed fishing location for a given angler group during an effort count determines their angler type.}

\item{angler_type_kayak_pontoon}{string passed from params that controls whether a boat designated as a kayak, pontoon, or kick during an effort count or angler group interview should be designated as a boat or bank angler.}
}
\value{
??
}
\description{
Dev note: is study_design option supposed to be "standard" or "tandard"? Or are we using str_detect to support multiple capitalizations?
}
