% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_inputs_bss.R
\name{prep_inputs_bss}
\alias{prep_inputs_bss}
\title{Create object/list of data inputs for BSS model}
\usage{
prep_inputs_bss(
  period,
  days,
  dwg_summarized,
  est_catch_group,
  census_expan,
  priors,
  study_design
)
}
\arguments{
\item{period}{?}

\item{days}{tibble with time strata and closure fields}

\item{dwg_summarized}{list with shared interview, index and census tibbles}

\item{est_catch_group}{data.frame passed from params of aggregated catch groups of interest to estimate}

\item{census_expan}{tibble summarizing p_census by angler_final and section_num where p_census is a hard coded value in the database specifying the proportion of a a section that is covered during a census count; values less than 1 with result in census counts being expanded (e.g., census count divided by p_census)}

\item{priors}{list of hyperpriors specified by user}

\item{study_design}{string passed from params denoting which study design was followed during data collection}
}
\value{
?
}
\description{
Create object/list of data inputs for BSS model
}
