% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_inputs_pe_daily_cpue_catch_est.R
\name{prep_inputs_pe_daily_cpue_catch_est}
\alias{prep_inputs_pe_daily_cpue_catch_est}
\title{Calculate daily cpue by section, date (period, day_type), angler_final, and catch group (aka est_cg) using the Ratio-of-Means (rom) calculation}
\usage{
prep_inputs_pe_daily_cpue_catch_est(
  days,
  dwg_summarized,
  angler_hours_daily_mean
)
}
\arguments{
\item{days}{tibble with time strata and closure fields}

\item{dwg_summarized}{list with shared interview, index and census tibbles}

\item{angler_hours_daily_mean}{tibble of daily mean angler hours by event_date, angler_type, and section_num

NOTE: the Ratio-of-Mean cpue calculation inherently weights angler groups with longer fishing times as opposed to Mean-of-Ratio (mor)
calculation which gives equal weight to every interview}
}
\value{
??
}
\description{
Calculate daily cpue by section, date (period, day_type), angler_final, and catch group (aka est_cg) using the Ratio-of-Means (rom) calculation
}
