% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_inputs_pe_days_total.R
\name{prep_inputs_pe_days_total}
\alias{prep_inputs_pe_days_total}
\title{Calculate total days the fishery was open per strata (strata = period, day_type, and section_num)}
\usage{
prep_inputs_pe_days_total(days)
}
\arguments{
\item{days}{tibble of dates and corresponding fields (e.g., year, month, period, open/close fishery by section)}
}
\value{
?
}
\description{
Calculate total days the fishery was open per strata (strata = period, day_type, and section_num)
}
