% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_inputs_pe_df.R
\name{prep_inputs_pe_df}
\alias{prep_inputs_pe_df}
\title{Calculate degrees of freedom based on number of creel survey days}
\usage{
prep_inputs_pe_df(angler_hours_daily_mean)
}
\arguments{
\item{angler_hours_daily_mean}{??}
}
\value{
??
}
\description{
Calculate degrees of freedom based on number of creel survey days
}
