% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_inputs_pe_int_ang_per_object.R
\name{prep_inputs_pe_int_ang_per_object}
\alias{prep_inputs_pe_int_ang_per_object}
\title{calculate the number of objects (e.g., trailers, vehicles, boats) each angler group brought to the fishery based on study_design used}
\usage{
prep_inputs_pe_int_ang_per_object(dwg_summarized, study_design)
}
\arguments{
\item{dwg_summarized}{list with shared interview, index and census tibbles}

\item{study_design}{parameter specifying study design and which if/else loop gets called below}

\item{...}{}
}
\value{
??
}
\description{
Devnote: we use summarize in the below. I assume it should be dplyr:: summarize?
}
