% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_inputs_pe_paired_census_index_counts.R
\name{prep_inputs_pe_paired_census_index_counts}
\alias{prep_inputs_pe_paired_census_index_counts}
\title{Summarize paired census and index angler effort counts}
\usage{
prep_inputs_pe_paired_census_index_counts(
  days,
  dwg_summarized,
  interview_ang_per_object,
  census_expan,
  study_design
)
}
\arguments{
\item{days}{tibble with time strata and closure fields}

\item{dwg_summarized}{list with shared interview, index and census tibbles}

\item{interview_ang_per_object}{tibble of interview-based values to translate vehicle/trailer counts to boat/bank}

\item{census_expan}{tibble summarizing p_census by angler_final and section_num
where p_census is a hard coded value in the database specifying the proportion of
a section that is covered during a census count; values less than 1 with result
in census counts being expanded (e.g., census count divided by p_census)}

\item{study_design}{parameter specifying study design and which if/else loop gets called below}
}
\value{
?
}
\description{
Devnote: uncertain about replace_na(1) -- is it doing as intended?
}
