% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrity.R
\name{search_fishery_name}
\alias{search_fishery_name}
\title{Find fishery names from a partial names}
\usage{
search_fishery_name(fishery_partial)
}
\arguments{
\item{fishery_partial}{Partial fishery name. Can take regular expressions, ignores capitalization.}
}
\value{
character vector.
}
\description{
Helps identify options for \code{fetch_dwg()}.
}
\examples{
search_fishery_name("gamefish")

}
