% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db_tables.R
\name{write_db_tables}
\alias{write_db_tables}
\title{Title}
\usage{
write_db_tables(con, analysis_lut, creel_estimates_db)
}
\arguments{
\item{con}{??}

\item{analysis_lut}{??}

\item{creel_estimates_db}{??}
}
\value{
?
}
\description{
This script validates that tables are ready to uploading and then defines functions to write to each table.
Currently write_lut includes robust error handling to ensure all fields write to the database, but this could be expanded.
}
\details{
Dev notes: defining functions within function. We may want to pull those out.
Dev notes: using assign. Not wild about that.
}
