% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_db_upload.R
\name{confirm_db_upload}
\alias{confirm_db_upload}
\title{Confirm database upload}
\usage{
confirm_db_upload(con, analysis_lut)
}
\arguments{
\item{con}{Connection to WDFW PostgreSQL database made with DBI-compliant RPostgres package. 'con' made with with establish_db_con.R function.}

\item{analysis_lut}{Data frame containing session-specific analysis_id and associated metadata created by generate_analysis_lut.R}
}
\value{
nothing returned.
}
\description{
This function attempts to verify that the model estimates have been written as expected and that the same number of
rows, columns, and values have been written. Currently this is in draft form and needs to be more comprehensive.
}
