% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_db_table.R
\name{fetch_db_table}
\alias{fetch_db_table}
\title{Query database tables}
\usage{
fetch_db_table(con, schema, table, filter = NULL)
}
\arguments{
\item{con}{Connection to WDFW PostgreSQL database made with DBI-compliant RPostgres package. 'con' created by establish_db_con.R function.}

\item{schema}{query schema}

\item{table}{table in database}

\item{filter}{filter to apply during query}
}
\value{
tibble of query results
}
\description{
Helper function to streamline queries of the WDFW Postgres database.
}
\examples{
\dontrun{
a <- fetch_db_table(con, "creel", "fishery_location_lut")
b <- fetch_db_table(con, "creel", "fishery_location_lut", filter = "survey_type == 'Index'")
c <- fetch_db_table(con, "creel", "fishery_location_lut",
filter = c("survey_type == 'Index'", "section_num == '1'"))
}
}
