% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bss_catch_daily.R
\name{get_bss_catch_daily}
\alias{get_bss_catch_daily}
\title{Get bss daily catch}
\usage{
get_bss_catch_daily(bss_fit, ecg, dwg)
}
\arguments{
\item{bss_fit}{??}

\item{ecg}{??}

\item{dwg}{description}
}
\value{
??
}
\description{
Get bss daily catch
}
