% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bss_effort_daily.R
\name{get_bss_effort_daily}
\alias{get_bss_effort_daily}
\title{Get bss effort}
\usage{
get_bss_effort_daily(bss_fit, ecg, ...)
}
\arguments{
\item{bss_fit}{??}

\item{ecg}{??}

\item{...}{??}
}
\value{
??
}
\description{
Get bss effort
}
