% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_est_pe_catch.R
\name{plot_est_pe_catch}
\alias{plot_est_pe_catch}
\title{plot PE catch estimates by catch group (est_cg), period (time stratum) and section}
\usage{
plot_est_pe_catch(estimates_pe_catch, est_catch_group, period_pe, estimates_pe)
}
\arguments{
\item{estimates_pe_catch}{??}

\item{est_catch_group}{??}

\item{period_pe}{??}

\item{estimates_pe}{??}
}
\value{
??
}
\description{
Dev note: is estimates_pe$catch supposed to be estimates_pe_catch in all the code below? Seems like it should be.
}
